+++ ribs/archives/_archive_base.py
@@ -6,7 +6,7 @@ import numpy as np
 from numpy_groupies import aggregate_nb as aggregate
 
 from ribs._utils import (check_1d_shape, check_batch_shape, check_finite,
-                         check_is_1d, check_solution_batch_dim)
+                         check_is_1d, validate_batch_args, validate_single_args)
 from ribs.archives._archive_data_frame import ArchiveDataFrame
 from ribs.archives._archive_stats import ArchiveStats
 from ribs.archives._cqd_score_result import CQDScoreResult
@@ -292,6 +292,7 @@ class ArchiveBase(ABC):  # pylint: disable = too-many-instance-attributes
             num_elites=0,
             coverage=self.dtype(0.0),
             qd_score=self.dtype(0.0),
+            norm_qd_score=self.dtype(0.0),
             obj_max=None,
             obj_mean=None,
         )
@@ -415,51 +416,6 @@ class ArchiveBase(ABC):  # pylint: disable = too-many-instance-attributes
         check_finite(measures, "measures")
         return self.index_of(measures[None])[0]
 
-    def _validate_add_args(self, solution_batch, objective_batch,
-                           measures_batch, metadata_batch):
-        """Performs preprocessing and checks for arguments to add()."""
-        solution_batch = np.asarray(solution_batch)
-        check_batch_shape(solution_batch, "solution_batch", self.solution_dim,
-                          "solution_dim", _ADD_WARNING)
-        batch_size = solution_batch.shape[0]
-
-        objective_batch = np.asarray(objective_batch, self.dtype)
-        check_is_1d(objective_batch, "objective_batch", _ADD_WARNING)
-        check_solution_batch_dim(objective_batch,
-                                 "objective_batch",
-                                 batch_size,
-                                 is_1d=True,
-                                 extra_msg=_ADD_WARNING)
-        check_finite(objective_batch, "objective_batch")
-
-        measures_batch = np.asarray(measures_batch)
-        check_batch_shape(measures_batch, "measures_batch", self.measure_dim,
-                          "measure_dim", _ADD_WARNING)
-        check_solution_batch_dim(measures_batch,
-                                 "measures_batch",
-                                 batch_size,
-                                 is_1d=False,
-                                 extra_msg=_ADD_WARNING)
-        check_finite(measures_batch, "measures_batch")
-
-        metadata_batch = (np.empty(batch_size, dtype=object) if
-                          metadata_batch is None else np.asarray(metadata_batch,
-                                                                 dtype=object))
-        check_is_1d(metadata_batch, "metadata_batch", _ADD_WARNING)
-        check_solution_batch_dim(metadata_batch,
-                                 "metadata_batch",
-                                 batch_size,
-                                 is_1d=True,
-                                 extra_msg=_ADD_WARNING)
-
-        return (
-            batch_size,
-            solution_batch,
-            objective_batch,
-            measures_batch,
-            metadata_batch,
-        )
-
     def add(self,
             solution_batch,
             objective_batch,
@@ -564,18 +520,22 @@ class ArchiveBase(ABC):  # pylint: disable = too-many-instance-attributes
         """
         self._state["add"] += 1
 
+        ## Step 0: Preprocess input. ##
+        solution_batch = np.asarray(solution_batch)
+        objective_batch = np.asarray(objective_batch)
+        measures_batch = np.asarray(measures_batch)
+        batch_size = solution_batch.shape[0]
+        metadata_batch = (np.empty(batch_size, dtype=object) if
+                          metadata_batch is None else np.asarray(metadata_batch,
+                                                                 dtype=object))
+
         ## Step 1: Validate input. ##
-        (
-            batch_size,
-            solution_batch,
-            objective_batch,
-            measures_batch,
-            metadata_batch,
-        ) = self._validate_add_args(
-            solution_batch,
-            objective_batch,
-            measures_batch,
-            metadata_batch,
+        validate_batch_args(
+            archive=self,
+            solution_batch=solution_batch,
+            objective_batch=objective_batch,
+            measures_batch=measures_batch,
+            metadata_batch=metadata_batch,
         )
 
         ## Step 2: Compute status_batch and value_batch ##
@@ -716,26 +676,12 @@ class ArchiveBase(ABC):  # pylint: disable = too-many-instance-attributes
             num_elites=len(self),
             coverage=self.dtype(len(self) / self.cells),
             qd_score=new_qd_score,
+            norm_qd_score=self.dtype(new_qd_score / self.cells),
             obj_max=new_obj_max,
             obj_mean=new_qd_score / self.dtype(len(self)),
         )
 
         return status_batch, value_batch
-
-    def _validate_add_single_args(self, solution, objective, measures,
-                                  metadata):
-        """Performs preprocessing and checks for arguments to add_single()."""
-        solution = np.asarray(solution)
-        check_1d_shape(solution, "solution", self.solution_dim, "solution_dim")
-
-        objective = self.dtype(objective)
-        check_finite(objective, "objective")
-
-        measures = np.asarray(measures)
-        check_1d_shape(measures, "measures", self.measure_dim, "measure_dim")
-        check_finite(measures, "measures")
-
-        return solution, objective, measures, metadata
 
     def add_single(self, solution, objective, measures, metadata=None):
         """Inserts a single solution into the archive.
@@ -776,16 +722,14 @@ class ArchiveBase(ABC):  # pylint: disable = too-many-instance-attributes
         """
         self._state["add"] += 1
 
-        (
-            solution,
-            objective,
-            measures,
-            metadata,
-        ) = self._validate_add_single_args(
-            solution,
-            objective,
-            measures,
-            metadata,
+        solution = np.asarray(solution)
+        objective = self.dtype(objective)
+        measures = np.asarray(measures)
+        validate_single_args(
+            self,
+            solution=solution,
+            objective=objective,
+            measures=measures,
         )
 
         index = self.index_of_single(measures)
@@ -853,6 +797,7 @@ class ArchiveBase(ABC):  # pylint: disable = too-many-instance-attributes
                 num_elites=len(self),
                 coverage=self.dtype(len(self) / self.cells),
                 qd_score=new_qd_score,
+                norm_qd_score=self.dtype(new_qd_score / self.cells),
                 obj_max=new_obj_max,
                 obj_mean=new_qd_score / self.dtype(len(self)),
             )
@@ -1172,7 +1117,8 @@ class ArchiveBase(ABC):  # pylint: disable = too-many-instance-attributes
                 size=(iterations, target_points, self.measure_dim),
             )
         else:
-            target_points = np.copy(target_points)  # Copy since we return this.
+            # Copy since we return this.
+            target_points = np.copy(target_points)
             if (target_points.ndim != 3 or
                     target_points.shape[0] != iterations or
                     target_points.shape[2] != self.measure_dim):