+++ ribs/archives/_archive_stats.py
@@ -1,10 +1,11 @@
 """Provides ArchiveStats."""
-from typing import NamedTuple
+import dataclasses
 
 import numpy as np
 
 
-class ArchiveStats(NamedTuple):
+@dataclasses.dataclass
+class ArchiveStats:
     """Holds statistics about an archive.
 
     Attributes of type :class:`~numpy.floating` will match the
@@ -21,6 +22,10 @@ class ArchiveStats(NamedTuple):
     #: QD score, i.e. sum of objective values of all elites in the archive.
     #: **This score only makes sense if objective values are non-negative.**
     qd_score: np.floating
+
+    #: Normalized QD score, i.e. the QD score divided by the number of cells in
+    #: the archive.
+    norm_qd_score: np.floating
 
     #: Maximum objective value of the elites in the archive. None if there are
     #: no elites in the archive.