+++ ribs/emitters/_emitter_base.py
@@ -1,5 +1,5 @@
 """Provides EmitterBase."""
-from abc import ABC, abstractmethod
+from abc import ABC
 
 import numpy as np